[staging] implement HTTP / REST proxy for the messaging broker
